% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createSQLiteDB.R
\name{createSQLiteDB}
\alias{createSQLiteDB}
\title{create SQLite database from maxquant txt results}
\usage{
createSQLiteDB(path2MQ, sqliteLocation = NULL, sqlname = NULL)
}
\arguments{
\item{path2MQ}{path to MaxQuant combined/txt folder}

\item{sqliteLocation}{where to write the sqliteDB (folder)}

\item{sqlname}{name of database file}
}
\description{
create SQLite database from maxquant txt results
}

