% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadFasta.R
\name{loadContaminantsNoMSQC1Fasta}
\alias{loadContaminantsNoMSQC1Fasta}
\title{get set difference of msqc1 proteins and fgcz contaminants}
\usage{
loadContaminantsNoMSQC1Fasta()
}
\description{
get set difference of msqc1 proteins and fgcz contaminants
}
\examples{
loadContaminantsNoMSQC1Fasta()[[1]]
}

