% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadMSQC1Transitions.R
\name{loadMSQC1Transitions}
\alias{loadMSQC1Transitions}
\title{load table with transitions}
\usage{
loadMSQC1Transitions()
}
\description{
load table with transitions
}
\examples{
loadMSQC1Transitions()
}

