% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadTheoryRatios.R
\name{loadTheoryHLRatios}
\alias{loadTheoryHLRatios}
\title{load theoretical Heavy Light ratios}
\usage{
loadTheoryHLRatios()
}
\description{
load theoretical Heavy Light ratios
}
\examples{
(loadTheoryHLRatios())
}

