% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadescription.R
\docType{package}
\name{p1755}
\alias{p1755}
\alias{p1755-package}
\title{Streamline descriptive analysis of quantitative data matrices}
\description{
Streamline descriptive analysis of quantitative data matrices
}

